API/requirements.txt
fastapi==0.111.0
numpy==1.24.3
Pillow==10.3.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
